refactor(#159): extract agent remediation routes to routers/remediation.py
